Fix up privacy, add documentation, and make a couple of improvements to various nits
